# src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: API de Gestión de Tareas
  description: API para gestionar tareas, usuarios y estados de tareas
  version: 1.0.0
  contact:
    name: Task Management API Support
    url: https://www.taskmanagement.com
    email: contacto@taskmanagement.com
  license:
    name: MIT License
    url: https://choosealicense.com/licenses/mit/
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/tareas:
    get:
      tags:
        - Tareas
      summary: Obtener todas las tareas
      description: Retorna una lista de todas las tareas disponibles
      operationId: ObtenerTareas
      responses:
        '200':
          description: Lista de tareas encontrada exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TareaResponseDTO'
    post:
      tags:
        - Tareas
      summary: Crear una nueva tarea
      description: Crea una nueva tarea con la información proporcionada
      operationId: crearTarea
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TareaDTO'
      responses:
        '201':
          description: Tarea creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TareaResponseDTO'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Usuario o estado de tarea no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/tareas/{id}:
    get:
      tags:
        - Tareas
      summary: Obtener una tarea por ID
      description: Retorna una tarea específica por su ID
      operationId: getTareaById
      parameters:
        - name: id
          in: path
          description: ID de la tarea
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Tarea encontrada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TareaResponseDTO'
        '404':
          description: Tarea no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Tareas
      summary: Actualizar una tarea
      description: Actualiza una tarea existente por su ID
      operationId: actualizarTarea
      parameters:
        - name: id
          in: path
          description: ID de la tarea a actualizar
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TareaDTO'
      responses:
        '200':
          description: Tarea actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TareaResponseDTO'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tarea, usuario o estado de tarea no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Tareas
      summary: Eliminar una tarea
      description: Elimina una tarea por su ID
      operationId: eliminarTarea
      parameters:
        - name: id
          in: path
          description: ID de la tarea a eliminar
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Tarea eliminada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TareaResponseDTO'
        '404':
          description: Tarea no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token de autenticación
  schemas:
    EstadoTarea:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID único del estado de tarea
          example: 1
          readOnly: true
        nombreEstado:
          type: string
          description: Nombre del estado de la tarea
          example: "En progreso"
      required:
        - nombreEstado
    TareaDTO:
      type: object
      properties:
        titulo:
          type: string
          description: Título de la tarea
          example: "Implementar autenticación"
        descripcion:
          type: string
          description: Descripción detallada de la tarea
          example: "Implementar sistema de autenticación usando JWT"
        estadoTareaId:
          type: integer
          format: int64
          description: ID del estado de la tarea
          example: 1
        usuarioId:
          type: integer
          format: int64
          description: ID del usuario asignado a la tarea
          example: 1
      required:
        - titulo
        - descripcion
        - estadoTareaId
        - usuarioId
    TareaResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID único de la tarea
          example: 1
        titulo:
          type: string
          description: Título de la tarea
          example: "Implementar autenticación"
        descripcion:
          type: string
          description: Descripción detallada de la tarea
          example: "Implementar sistema de autenticación usando JWT"
        estadoTarea:
          $ref: '#/components/schemas/EstadoTarea'
        usuario:
          $ref: '#/components/schemas/UsuarioResponseDTO'
...
    UsuarioResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID único del usuario
          example: 1
        nombreUsuario:
          type: string
          description: Nombre de usuario
          example: "john.perez"
        correo:
          type: string
          format: email
          description: Correo electrónico del usuario
          example: "john.perez@example.com"
    ErrorResponse:
      type: object
      additionalProperties:
        type: string
      description: Objeto que contiene los mensajes de error
      example:
        "Mensaje": "Recurso no encontrado"
        "Status": "404 NOT_FOUND"
...